Lowered, Raised and APK endpoints

+++ new_backend/src/main/resources/db/V01__create_products_table.sql
@@ -81,5 +81,5 @@ CREATE TABLE products (
     "updatedAt" TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     "oldPrice" NUMERIC(10,2) DEFAULT NULL,
     "dateOfPriceChange" TIMESTAMP DEFAULT NULL,
-    "apk" NUMERIC(10, 2)
+    "apk" DOUBLE PRECISION
 );
